refactor: moving the diagnostics settings category func into the main file
